we will make another container in recipe-card. we'll name the class "container01".Now we gave to add boxes in the container,so we will use div element with class to make differnt 'empty boxes'.we'll name them box1, box2,.....
For now they may be not visible ,so we'll add
border to them by using .box selectors to target
the all boxes.
in box1,box2,....... selector well use border
property to add color and to make the visible
we'll set the to solid.At last add height and
width.
before the boxes we also added a heading .