now we will add names to all the iboxes by writing their names in <p>.after that we want the names to act as a link to he particular manga. for that we will use <a> iside of<p>, after that inside <a> i.e. href we'll add the url of the any other website ,now you can also use rel to target the page to open in a new tab, but for now we will use the same tab so no need of rel.
we want are header links to be active, so we'll add
<a> to them .so now we want to jump on a particular
place on our page so we'll add href in <a>, for the
jump to work, we'll have to add ID to the part we
want to jump , so we'll add id to all the headings
we want to jump. now in href use # with the id name
to jump to that site of our page.